Stop parsing / as a command line switch
I had added / as a switch to cover the case of /? which is a common way
to see usage info in Windows console programs. However MinGW built
programs will attempt to glob that argument as paths.

For example, consider two single letter directories at the root of the
current drive, X:\A and X:\B. Attempting to run a mingw-built program
with /? will change that argument to two arguments /A and /B which is
clearly not what the user intended.