Check plugged-in status for change
This commit is to bring the status changes more in line with the Windows
battery systray tooltip. For example when a laptop is plugged in it
won't necessarily go to charging or have its percentage remaining
changed, but the systray battery tooltip contents may still change.

Prior to this commit, when only plugged-in status changed during default
monitor mode the status shown would not be updated until a change in
battery presence, charging status or percent remaining.

Verbose monitor mode already checks ACLineStatus (on which plugged-in
status is based) and is not affected by this change.