Process window messages more often
Prior to this change window messages were only processed in the loop
before getting the power status. This commit changes it so that window
messages are also processed after getting the power status, and during
the wait when the power status fails.

This commit is mainly to lessen the chance that all related window
messages for a power status change are not processed (and shown) before
showing that status (if it has changed). The OS does not guarantee that
all related window messages are sent to the monitor window before
updating the power status, therefore there's no way to be certain.

Before:
[Wed Jun 28 03:21:50 PM]: Fully charged (100%)
[Wed Jun 28 03:21:50 PM]: WM_POWERBROADCAST: PBT_APMPOWERSTATUSCHANGE

After (this order not guaranteed, but more likely):
[Wed Jun 28 03:21:50 PM]: WM_POWERBROADCAST: PBT_APMPOWERSTATUSCHANGE
[Wed Jun 28 03:21:50 PM]: Fully charged (100%)